Moved collect functions to helper file
